Refactor test setup: extract server into separate struct
